+++ src/snowflake/snowpark/mock/_util.py
@@ -2,6 +2,7 @@
 # Copyright (c) 2012-2025 Snowflake Computing Inc. All rights reserved.
 #
 
+from decimal import ROUND_HALF_UP, Context
 import math
 import os
 import shutil
@@ -19,6 +20,7 @@ from snowflake.snowpark.types import (
     BooleanType,
     ByteType,
     DateType,
+    DecFloatType,
     DecimalType,
     DoubleType,
     FloatType,
@@ -244,6 +246,7 @@ def fix_drift_between_column_sf_type_and_dtype(col: ColumnEmulator):
         BooleanType: bool,
         ByteType: numpy.int8 if not col.sf_type.nullable else "Int8",
         DateType: object,
+        DecFloatType: numpy.float64,
         DecimalType: numpy.float64,
         DoubleType: numpy.float64,
         FloatType: numpy.float64,
@@ -332,3 +335,13 @@ class ImportContext:
 
     def __exit__(self, type, value, traceback):
         self._callback()
+
+
+# context to instantiate decimal.Decimal values comforming to SF DECFLOAT data type
+# https://docs.snowflake.com/en/sql-reference/data-types-numeric#label-data-type-decfloat
+DECFLOAT_CONTEXT = Context(
+    prec=38,
+    Emin=-16383,
+    Emax=16384,
+    rounding=ROUND_HALF_UP,
+)